refactor: use RdbEngineFactory.create instead of RdbEngineStrategy.create
