+++ core/src/test/java/com/scalar/db/storage/jdbc/JdbcAdminTestBase.java
@@ -55,9 +55,9 @@ public abstract class JdbcAdminTestBase {
 
   private JdbcAdmin createJdbcAdminFor(RdbEngine rdbEngine) {
     // Arrange
-    RdbEngineStrategy st = RdbEngineStrategy.create(rdbEngine);
+    RdbEngineStrategy st = RdbEngineFactory.create(rdbEngine);
     try (MockedStatic<RdbEngineStrategy> mocked = mockStatic(RdbEngineStrategy.class)) {
-      mocked.when(() -> RdbEngineStrategy.create(any(JdbcConfig.class))).thenReturn(st);
+      mocked.when(() -> RdbEngineFactory.create(any(JdbcConfig.class))).thenReturn(st);
       return new JdbcAdmin(dataSource, config);
     }
   }
